Beginnings of content-extraction ML training
I see no reason not to use the `dragnet` library to do some heavy lifting for us. We'll still be able to tune things plenty (e.g. pick our own sklearn learning models), and it's quick to spin up.

This commit provides:
- `train_model.py`: A function designed for spinning up and training a new model using dragnet
- `content_data/`: A directory of HTML files and "corrected" text files (corrected, to dragnet, means the text we expect to see returned by our model for a given HTML document). Right now, this contains only data from one website, but I'm actively working on that
- Updates to .gitignore to block out some dragnet output; while I don't think it'd be horrible to commit, I'd like to be more intentional with what we save (i.e. we'll make a directory in the future for any model(s) we'd like to save)

See the dragnet repository on Github for more information about what this all is doing, but it's relatively straightforward.